Safari-609.1.11.1/LayoutTests/accessibility/gtk/aria-modal-state-exposed-expected.txt

This tests that a dialog with aria-modal='true' exposes ATK_STATE_MODAL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axElement.boolAttributeValue('AXModal') is true
PASS axElement.boolAttributeValue('AXModal') is false
PASS successfullyParsed is true

TEST COMPLETE
